requirements.txt
protobuf
grpcio
grpcio-tools
boto3
watchtower
setuptools
wheel
python_daemon
psutil